\name{plot.bacr}
\alias{plot.bacr}
\title{
Plot Method for Class 'bacr'
}
\description{
Plot the posterior inclusion probabilities of potential confounders for a \code{bacr} object
}
\usage{
\method{plot}{bacr}(x, \dots)
}
\arguments{
  \item{x}{
An object returned by the bac function
}
\item{\dots}{
arguments passed to or from other methods.
}
}
\author{
Chi Wang
}
\seealso{
 \code{\link{bac}}, \code{\link{summary.bacr}}
}
